+++ packages/vuikit/src/drop/index.js
@@ -1,6 +1,7 @@
 import * as constants from './constants'
-import { active } from './drop/toggle'
 
-export { constants, active }
 export * from './elements'
 export * from './components'
+
+export { constants }
+export { active } from './instances'