+++ packages/vuikit/src/dropdown/components/index.js
@@ -1,0 +1,2 @@
+export { default as Dropdown } from './dropdown'
+export { default as DropdownNav } from './dropdown-nav'