+++ packages/vuikit/src/dropdown/elements/index.js
@@ -1,0 +1,2 @@
+export { default as ElDropdown } from './dropdown'
+export { default as ElDropdownNav } from './dropdown-nav'