+++ packages/vuikit/src/dropdown/index.js
@@ -1,1 +1,2 @@
+export * from './elements'
 export * from './components'