+++ packages/vuikit/src/nav/components/index.js
@@ -1,7 +1,6 @@
 export { default as NavItem } from './nav--item'
 
 export { ElNav as Nav } from '../elements'
-export { ElNavDropdown as NavDropdown } from '../elements'
 export { ElNavItemDivider as NavItemDivider } from '../elements'
 export { ElNavItemHeader as NavItemHeader } from '../elements'
 export { ElNavItemParent as NavItemParent } from '../elements'