+++ packages/vuikit/src/subnav/components/index.js
@@ -1,3 +1,4 @@
 export { default as Subnav } from './subnav'
 export { default as SubnavItem } from './subnav--item'
 export { default as SubnavItemDropdown } from './subnav--item-dropdown'
+export { default as SubnavItemDropdownNav } from './subnav--item-dropdown-nav'